requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
numpy==2.1.1